pyproject.toml
[project]
name = "student360-ai"
version = "0.1.0"
description = "Multi-agent AI service for Student360"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # LangGraph + LangChain
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",

    # Google Gemini
    "google-genai>=0.1.0",

    # Database
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # Redis + Queue
    "redis>=5.0.0",
    "arq>=0.26.0",

    # HTTP client
    "httpx>=0.27.0",

    # Logging
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",   # for TestClient
]
phase-b = [
    # CV parsing (Phase B)
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
]
phase-d = [
    # Receipt OCR (Phase D)
    "google-cloud-vision>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
